- attempt compilation fix for mingw

+++ src/sbml/math/FormulaParser.cpp
@@ -398,7 +398,7 @@ LIBSBML_EXTERN
 ASTNode_t *
 SBML_parseFormula (const char *formula)
 {
-  long rule, state, action;
+  long long rule, state, action;
 
   ASTNode_t *node = NULL;
 
@@ -417,7 +417,7 @@ SBML_parseFormula (const char *formula)
 
   while (1)
   {
-    state  = (long) Stack_peek(stack);
+    state  = (long long) Stack_peek(stack);
     action = FormulaParser_getAction(state, token);
 
     if (action == ACCEPT_STATE)
